requirements.txt
absl-py==2.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
fastapi==0.135.3
gymnasium==1.2.3
httpx==0.28.1
llama-cpp-python==0.3.20
fonttools==4.62.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
idna==3.11
immutabledict==4.3.1
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
llvmlite==0.47.0
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.19.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
nvidia-nccl-cu12==2.29.7
ortools==9.15.6755
packaging==26.0
pandas==3.0.2
pillow==12.2.0
plotly==6.7.0
protobuf==6.33.6
pyarrow==23.0.1
pydantic==2.13.0
pydantic_core==2.46.0
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
shap==0.51.0
stable-baselines3==2.8.0
six==1.17.0
slicer==0.0.8
smmap==5.0.3
starlette==1.0.0
streamlit==1.56.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
watchdog==6.0.0
xgboost==3.2.0
